requirements.txt
fastapi==0.115.12
huggingface_hub==0.31.1
nemo_toolkit[asr]==2.4.0
numpy==1.26.4
orjson==3.10.16
pydantic-settings==2.9.1
python-multipart==0.0.20
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
uvicorn[standard]==0.34.2
